Enhancement Order

+++ ElbayanServices/Repository/Customers/DeferredPayments/Dtos/OrderNameDto.cs
@@ -10,7 +10,7 @@ namespace ElbayanServices.Repository.Customers.DeferredPayments.Dtos
  public class OrderNameDto
     {
         public Guid Id { get; set; }
-        public int OrderNumber { get; set; }
+        public long OrderNumber { get; set; }
 
     }
 }
